examples/storytelling_agent/requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.6
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.29
langchain>=0.1.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0 